REF: PseudoSingle is a SoftPositioner
